Implement Jacob's termination criterion for radial sweep.
Each pixel can be visited multiple times, but only once from each direction. Visit directions are memorized.

There is a problem with a proper termination criterion. Currently the lines get unacceptably long!
We'd like to allow multiple visits in order to reduce the number of lone pixel, but only in a limited way - in special situations where it pays off.